DJANGO/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.11
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
fake-useragent==2.0.3
fonttools==4.55.3
gunicorn==22.0.0
h11==0.16.0
idna==3.10
Markdown==3.7
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.1
stripe==12.0.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.2
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
wheel==0.38.1
whitenoise==6.6.0
zopfli==0.2.3.post1
redis==6.2.0
async-timeout==5.0.1
channels_redis==4.2.1